Migrate to numeric

+++ api/lib/schema.ts
@@ -5,7 +5,7 @@ import { geometry, GeometryType } from '@openaddresses/batch-generic';
 import { ConnectionAuth } from './connection-config.js';
 import { TAKGroup, TAKRole } from  './api/types.js';
 import { Layer_Priority, Profile_Stale, Profile_Speed, Profile_Elevation, Profile_Distance } from  './enums.js';
-import { json, boolean, integer, timestamp, pgTable, serial, varchar, text, unique, index } from 'drizzle-orm/pg-core';
+import { json, boolean, numeric, integer, timestamp, pgTable, serial, varchar, text, unique, index } from 'drizzle-orm/pg-core';
 
 /** Internal Tables for Postgis for use with drizzle-kit push:pg */
 export const SpatialRefSys = pgTable('spatial_ref_sys', {
@@ -322,7 +322,7 @@ export const ProfileOverlay = pgTable('profile_overlays', {
     updated: timestamp('updated', { withTimezone: true, mode: 'string' }).notNull().default(sql`Now()`),
     pos: integer('pos').notNull().default(5),
     type: text('type').notNull().default('vector'),
-    opacity: integer('opacity').notNull().default(1),
+    opacity: numeric('opacity').notNull().default(1),
     visible: boolean('visible').notNull().default(true),
     token: text('token'),
     styles: json('styles'),
